Simplify jsconfig.json removal

+++ compile_gui.py
@@ -12,10 +12,8 @@ shutil.copy('end_template.txt', 'compile_temp/')
 shutil.copytree('Library', 'compile_temp/Library')
 
 # Remove jsconfig file, which only exists on my machine
-try:
+if os.path.isfile('compile_temp/Library/jsconfig.json'):
     os.remove('compile_temp/Library/jsconfig.json')
-except FileNotFoundError:
-    pass
 
 # Run pyinstaller with correct flags from command prompt (I'm on Windows and haven't tested this on Linux or MacOS)
 os.system('cmd /c "pyinstaller formatter_gui.py -wF -n WhatsApp_Formatter --distpath ./compile_temp -i '
